Create abstract EmailSender and extract CurlEmailSender
